requirements.txt
django==4.0.3
django-model-utils==4.2.0
djangorestframework==3.13.1
requests==2.27.1
urllib3==1.26.8
xmltodict==0.12.0
